requirements.txt
Django>=1.3
surlex
sorl-thumbnail
adminsortable
